+++ src/FSharpx.Async/Async.fs
@@ -5,6 +5,7 @@
 namespace FSharpx.Control
 open System
 open System.Threading
+open System.Threading.Tasks
 
 // ----------------------------------------------------------------------------
 
@@ -60,6 +61,50 @@ module AsyncExtensions =
         let! c = c
         return a,b,c }
 
+      /// Creates an async computation which runs the provided sequence of computations and completes
+      /// when all computations in the sequence complete. Up to parallelism computations will
+      /// be in-flight at any given point in time. Error or cancellation of any computation in
+      /// the sequence causes the resulting computation to error or cancel, respectively.
+      static member ParallelIgnore (parallelism:int) (xs:seq<Async<_>>) = async {
+                
+        let sm = new SemaphoreSlim(parallelism)
+        let cde = new CountdownEvent(1)
+        let tcs = new TaskCompletionSource<unit>()
+    
+        let inline ok _ =
+          sm.Release() |> ignore
+          if (cde.Signal()) then
+            tcs.SetResult(())
+                
+        let inline err (ex:exn) =
+          tcs.SetException ex
+          sm.Release() |> ignore
+                
+        let inline cnc (ex:OperationCanceledException) =      
+          tcs.SetCanceled()
+          sm.Release() |> ignore
+
+        try
+
+          for computation in xs do
+            sm.Wait()
+            cde.AddCount(1)            
+            // the following decreases throughput 3x but avoids blocking
+            // do! sm.WaitAsync() |> Async.AwaitTask
+            Async.StartWithContinuations(computation, ok, err, cnc)                    
+      
+          if (cde.Signal()) then
+            tcs.SetResult(())
+
+          do! tcs.Task |> Async.AwaitTask
+
+        finally
+      
+          cde.Dispose()    
+          sm.Dispose()
+                          
+        }
+
       /// An async computation which does nothing.
       static member inline unit = AsyncOps.unit
 
@@ -71,3 +116,46 @@ module AsyncExtensions =
       /// async computation to the specified function. The computation produced 
       /// by the specified function is returned.
       static member inline bind f a = async.Bind(a, f)
+
+      /// Maps over an async computation which produces a choice value
+      /// using a function which maps over Choice1Of2 and itself returns a choice. 
+      /// A value of Choice2Of2 is treated like an error and passed through.
+      static member mapChoice (f:'a -> Choice<'b, 'e>) (a:Async<Choice<'a, 'e>>) : Async<Choice<'b, 'e>> = 
+        a |> Async.map (function 
+          | Choice1Of2 a' -> f a'
+          | Choice2Of2 e -> Choice2Of2 e)
+
+      /// Binds an async computation producing a choice value to another async
+      /// computation producing a choice such that a Choice2Of2 value is passed through.
+      static member bindChoice (f:'a -> Async<Choice<'b, 'e>>) (a:Async<Choice<'a, 'e>>) : Async<Choice<'b, 'e>> = 
+        a |> Async.bind (function
+          | Choice1Of2 a' -> f a'
+          | Choice2Of2 e ->  Choice2Of2 e |> async.Return)
+
+      /// Binds an async computation producing a choice value to another async
+      /// computation producing a choice such that a Choice2Of2 value is passed through.
+      static member bindChoices (f:'a -> Async<Choice<'b, 'e2>>) (a:Async<Choice<'a, 'e1>>) : Async<Choice<'b, Choice<'e1, 'e2>>> = 
+        a |> Async.bind (function
+          | Choice1Of2 a' -> f a' |> Async.map (function Choice1Of2 b -> Choice1Of2 b | Choice2Of2 e2 -> Choice2Of2 (Choice2Of2 e2))
+          | Choice2Of2 e1 -> Choice2Of2 (Choice1Of2 e1) |> async.Return)
+
+      /// Creates a computation which produces a tuple consiting of the value produces by the first
+      /// argument computation to complete and a handle to the other computation. The second computation
+      /// to complete is memoized.
+      static member internal chooseBoth (a:Async<'a>) (b:Async<'a>) : Async<'a * Async<'a>> =
+        Async.FromContinuations <| fun (ok,err,cnc) ->
+          let state = ref 0            
+          let tcs = TaskCompletionSource<'a>()            
+          let inline ok a =
+            if (Interlocked.CompareExchange(state, 1, 0) = 0) then
+              ok (a, tcs.Task |> Async.AwaitTask)
+            else
+              tcs.SetResult a
+          let inline err (ex:exn) =
+            if (Interlocked.CompareExchange(state, 1, 0) = 0) then err ex
+            else tcs.SetException ex
+          let inline cnc ex =
+            if (Interlocked.CompareExchange(state, 1, 0) = 0) then cnc ex
+            else tcs.SetCanceled()
+          Async.StartWithContinuations(a, ok, err, cnc)
+          Async.StartWithContinuations(b, ok, err, cnc)