+++ RELEASE_NOTES.md
@@ -1,5 +1,6 @@
 ### 1.11.1 - 26.03.2015
 * BUGFIX: AsyncSeq.toBlockingSeq does not hung forever if an exception is thrown and reraise it outside - https://github.com/fsprojects/FSharpx.Async/pull/21
+* Added Async.bindChoice, Async.ParallelIgnore, AsyncSeq.zipWithAsync, AsyncSeq.zappAsync, AsyncSeq.threadStateAsync, AsyncSeq.merge, AsyncSeq.traverseOptionAsync, AsyncSeq.traverseChoiceAsync
 
 ### 1.11.0 - 27.02.2015
 * Added Async.map, Async.bind, Async.unit